# backend/Document/Swagger.yaml
openapi: 3.0.4
info:
  title: National Football League Management API
  description: |-
    This is an API for managing a national football league, built on the OpenAPI 3.0 specification. It provides endpoints for managing team information, players, match schedules, results, and league standings.  
    Learn more about OpenAPI at [https://swagger.io](https://swagger.io). We follow a design-first approach, allowing you to contribute to improving the API, from its definition to its code.  
    Some useful links:  
    - [National Football League API Repository](https://github.com/truonghoangkhiem/QL_Giai_BD_QG_SE104_UIT.git)
  termsOfService: https://swagger.io/terms/
  contact:
    email: truongkhiemhoang@gmail.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.12
...
servers:
  - url: http://localhost:5000
...
paths:
...
  /api/player_rankings:
    post:
      tags:
        - playerranking
      summary: Create a new player ranking
      description: Creates a new player ranking for a specific season and player result
      operationId: createPlayerRankings
      security:
        - bearerAuth: []
      requestBody:
        description: Information needed to create a player ranking
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePlayerRankingRequest"
      responses:
        "201":
          description: Player ranking created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Invalid input or player ranking already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized (missing or invalid token)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Player result not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/player_rankings/season/{seasonid}:
    get:
      tags:
        - playerranking
      summary: Get player rankings by season ID and date
      description: Retrieves the latest player rankings for a specific season up to a given date
      operationId: getPlayerRankingsbySeasonIdAndDate
      parameters:
        - $ref: "#/components/parameters/SeasonId"
      requestBody:
        description: Date to filter player rankings
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FilterDateRequest"
      responses:
        "200":
          description: Latest player rankings retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerRankingsResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No player rankings found for this season and date
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/player_rankings/match/{matchid}:
    put:
      tags:
        - playerranking
      summary: Update player rankings after a match
      description: Updates player rankings for a specific season based on match results
      operationId: updatePlayerRankingsafterMatch
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/MatchId"
      responses:
        "200":
          description: Player rankings updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdatePlayerRankingsResponse"
        "400":
          description: Invalid match ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized (missing or invalid token)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Match or player results not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/player_rankings/{id}:
    delete:
      tags:
        - playerranking
      summary: Delete a player ranking
      description: Deletes an existing player ranking by its ID
      operationId: deletePlayerRankings
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PlayerRankingId"
      responses:
        "200":
          description: Player ranking deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Invalid player ranking ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized (missing or invalid token)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Player ranking not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    SeasonId:
      name: season_id
      in: path
      required: true
      schema:
        type: string
      example: "67d0ec1a412c1edc00d2117c"
...
    MatchId:
      name: id
      in: path
      required: true
      schema:
        type: string
      example: "67d0ec1a412c1edc00d2117b"
...
    PlayerRankingId:
      name: id
      in: path
      required: true
      schema:
        type: string
      example: "67d0ec1a412c1edc00d21182"
  schemas:
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          example: "error"
        message:
          type: string
          example: "Invalid email address"
      required:
        - status
        - message
    SuccessResponse:
      type: object
      properties:
        status:
          type: string
          example: "success"
        message:
          type: string
          example: "Operation successful"
        data:
          type: object
          nullable: true
          example: null
      required:
        - status
        - message
...
    PlayerRanking:
      type: object
      properties:
        _id:
          type: string
          example: "67d0ec1a412c1edc00d21182"
        player_result_id:
          type: string
          example: "67d0ec1a412c1edc00d21180"
        season_id:
          type: string
          example: "67d0ec1a412c1edc00d2117c"
        rank:
          type: integer
          example: 1
        date:
          type: string
          format: date-time
          example: "2025-04-08T00:00:00Z"
        createdAt:
          type: string
          format: date-time
          example: "2025-04-08T12:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2025-04-08T12:00:00Z"
      required:
        - _id
        - player_result_id
        - season_id
        - rank
        - date
        - createdAt
        - updatedAt
    PlayerRankingsResponse:
      type: object
      properties:
        status:
          type: string
          example: "success"
        message:
          type: string
          example: "Fetched player rankings successfully"
        data:
          type: array
          items:
            $ref: "#/components/schemas/PlayerRanking"
      required:
        - status
        - message
        - data
    CreatePlayerRankingRequest:
      type: object
      properties:
        player_result_id:
          type: string
          example: "67d0ec1a412c1edc00d21180"
        season_id:
          type: string
          example: "67d0ec1a412c1edc00d2117c"
        rank:
          type: integer
          example: 1
        date:
          type: string
          format: date-time
          example: "2025-04-08T00:00:00Z"
      required:
        - player_result_id
        - season_id
        - rank
        - date
    UpdatePlayerRankingsResponse:
      type: object
      properties:
        status:
          type: string
          example: "success"
        message:
          type: string
          example: "Player rankings updated successfully"
        data:
          type: object
          nullable: true
          example: null
      required:
        - status
        - message
    FilterDateRequest:
      type: object
      properties:
        date:
          type: string
          format: date-time
          example: "2025-04-08T00:00:00Z"
      required:
        - date